Updated to new Render System
Removed every usage of SDL so the user can quickly and easely switch to a new render system.

+++ RealEngine/ButtonPrefab.cpp
@@ -76,7 +76,7 @@ void real::ButtonPrefab::InitComponents(glm::ivec2 pos, glm::ivec2 size, std::sh
 	m_pGameObject->GetComponent<TransformComponent>()->SetLocalPosition(pos);
 	m_pGameObject->AddComponent<TextureComponent>()->SetRenderOffset(size.x / 2, size.y / 2);
 	m_pGameObject->AddComponent<ColliderComponent>(size)->EnableDebugRendering(true);
-	m_pGameObject->GetComponent<ColliderComponent>()->Translate(-(size.x / 2.f), -(size.y / 2.f));
+	m_pGameObject->GetComponent<ColliderComponent>()->Translate(-(static_cast<float>(size.x) / 2.f), -(static_cast<float>(size.y) / 2.f));
 
 	if (texture != nullptr)
 		m_pGameObject->GetComponent<TextureComponent>()->SetTexture(std::move(texture));
